Split apart the massive utils.py
