+++ pipenv/environment.py
@@ -12,6 +12,7 @@ from sysconfig import get_paths, get_python_version
 import pkg_resources
 
 import pipenv
+import utils.indexes
 
 from pipenv.environments import is_type_checking
 from pipenv.utils import make_posix, normalize_path, subprocess_run
@@ -598,7 +599,7 @@ class Environment:
         from .vendor.pip_shims.shims import InstallCommand, get_package_finder
 
         pip_command = InstallCommand()
-        pip_args = self._modules["pipenv"].utils.prepare_pip_source_args(self.sources)
+        pip_args = utils.indexes.prepare_pip_source_args(self.sources)
         pip_options, _ = pip_command.parser.parse_args(pip_args)
         pip_options.cache_dir = self.project.s.PIPENV_CACHE_DIR
         pip_options.pre = self.pipfile.get("pre", pre)
