+++ pipenv/utils/__init__.py
@@ -1,0 +1,13 @@
+import logging
+from .constants import *
+from .dependencies import *
+from .indexes import *
+from .internet import *
+from .locking import *
+from .processes import *
+from .resolver import *
+from .shell import *
+from .spinner import *
+from .toml import *
+
+logging.basicConfig(level=logging.ERROR)
